+++ lib/main.dart
@@ -1,65 +1,31 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'tutorial_1/animation.dart';
+//import './charch.dart';
 
-void main() => runApp(const LogoApp());
+void main() => runApp(MainApp());
 
-class AnimatedLogo extends AnimatedWidget {
-  const AnimatedLogo({Key? key, required Animation<double> animation})
-      : super(key: key, listenable: animation);
-
-  static final _opacityTween = Tween<double>(begin: 0.1, end: 1); //透明度
-  static final _sizeTween = Tween<double>(begin: 0, end: 300); //ロゴの大きさの範囲
-
+class MainApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final animation = listenable as Animation<double>;
-    return Center(
-      child: Opacity(
-        opacity: _opacityTween.evaluate(animation),
-        child: Container(
-          margin: const EdgeInsets.symmetric(vertical: 10),
-          height: _sizeTween.evaluate(animation),
-          width: _sizeTween.evaluate(animation),
-          child: const FlutterLogo(),
-        ),
-      ),
-    );
+    return MaterialApp(home: PageChanege());
   }
 }
 
-class LogoApp extends StatefulWidget {
-  const LogoApp({Key? key}) : super(key: key);
-
+class PageChanege extends StatelessWidget {
   @override
-  _LogoAppState createState() => _LogoAppState();
-}
-
-class _LogoAppState extends State<LogoApp> with SingleTickerProviderStateMixin {
-  late Animation<double> animation;
-  late AnimationController controller;
-
-  @override
-  void initState() {
-    super.initState();
-    controller = AnimationController(
-        duration: const Duration(seconds: 2), vsync: this); //2秒で変化させる
-    animation = CurvedAnimation(parent: controller, curve: Curves.easeIn)
-      ..addStatusListener((status) {
-        //アニメーションの状態が切り替わったときに呼び出される
-        if (status == AnimationStatus.completed) {
-          controller.reverse();
-        } else if (status == AnimationStatus.dismissed) {
-          controller.forward();
-        }
-      });
-    controller.forward();
-  }
-
-  @override
-  Widget build(BuildContext context) => AnimatedLogo(animation: animation);
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(title: Text('☆☆')),
+        body: Center(
+            child: ElevatedButton(
+                child: const Text("flutter_tutorial"),
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                        builder: (context) => const AnimationPage()),
+                  );
+                })));
   }
 }
