+++ lib/tutorial_1/charch.dart
@@ -1,10 +1,5 @@
 import 'package:flutter/material.dart';
 
-//アプリ開始
-void main() {
-  runApp(const MyApp());
-}
-
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
