+++ test/widget_test.dart
@@ -7,13 +7,14 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:tutorial/tutorial_1/animation.dart';
 
-import 'package:tutorial/main_2.dart';
+import 'package:tutorial/main.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+    await tester.pumpWidget(MainApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
